fix: notConnected behaviour after moving to wagmi 2.0
Prior to wagmi 2.0 `useChain` hook was used to determine if correct chain is connected as well as to check if wallet is connected. In wagmi 2.0 the hook was replaced with useAccount and behaviour slightly changed, with a correct way to detect unconnected wallet to check for isConnected property. And `chain` will can still be returned if user haven't connected his account to the dapp